Make pellets "round" via shaders
Also change variant to float, because uints are wierd in shaders

+++ include/glove/VertexFormats.h
@@ -18,5 +18,5 @@ struct Vertex2DTexRgbav {
 	glm::vec2 pos;     ///< Position
 	glm::vec2 uv;      ///< Normalized texture coordinates
 	glm::vec4 rgba;    ///< rgba OR xyzw
-	uint32_t  variant; ///< Is rgba or xyzw
+	float     variant; ///< Is rgba or xyzw
 };
